fix: merge configurations
